Switch from IPCClient to D-Bus
Create an helper class to handle the communication to the usbguard-dbus
daemon instead of using IPCClient to connect to the usbguard-daemon
directly. This is done because the usbguard library does not have a
stable API/ABI yet, while the D-Bus interface is easier and versioned.

The added DBusBridge also watches the usbguard D-Bus service to handle
its registration and unregistration, so there is no need to poll every
second when the service is not available.