Build the translations
Require the Qt linguist tools to build the translations, and add the
cmake rules for that. Since we want to embed them as resources, and not
hardcode the list of available files, create an helper script that
outputs a qrc file with all the available qm files.

Related change: update the TRANSLATIONS.md file with instructions.