Add a Rule class
Add a local Rule class that is supposed to usbguard::Rule, with the idea
to replace it; because of this
- the exposed API is what is used in the current codebase
- the Target enum, and the ImplicitID constant are taken from usbguard,
  as they must be precise values
- the API uses QString instead of std::string: this simplifies a bit
  the usage of Rule within the Qt codebase

Internally, usbguard::Rule is currently wrapped, as other usbguard
classes are still used; still, this eases a bit the future work of
- switching away from existing usbguard classes
- switch the internal Rule implementation to something else (e.g. a
  custom parser)

Also, add QDebug stream support for Rule and Rule::Target, making use
of it to reduce the need of converting them to string manually.

Despite all the changes, there should be hopefully no behaviour changes.