DeviceModel: fix device removal
removeDevice(quint32) calls removeDevice(DeviceModelItem*, bool),
and the 'notify' boolean of it was always set to false; since the
QAbstractItemModel APIs for index manipulation (e.g. beginRemoveRows())
was used only when 'notify' is true, this means the model did not
properly report the removal of items to other parties (e.g. the view and
the selection model).

Fix the removal of devices, so things work properly:
- when removing a device by ID, always notify
  we are removing the item altogether, so that should be enough to drop
  the whole subtree of the item
- delete the item before calling endRemoveRows(), otherwise after that
  notification the state is inconsistent