Drop extra selection model & view resets
Since DeviceModel properly notifies now when items are removed (either
one by one, or all of them by clearing the model), there is no need to
manually clear the selection model or the view, as both are kept
up-to-date.

This also fixes a small bug/annoyance: when an item is selected, and
a device that is not a parent of the selected one is removed, then the
currently selected item is kept selected as expected.