Do not parse rules unnecessarily when loading devices
The rule ID is available already from the list of rules loaded via
D-Bus; hence, parse a rule only if we know that a rule is already
present in the device model.

This avoids extra rule parsing, especially when loading the devices
after a device insertion event, and loading the added ones.