Explicitly use int as type
In Qt 5 QString::size() returns int, which is the same as the deduced
type for "1"; in Qt 6 QString::size() return qsizetype, which is ssize_t
(and thus long long on 64bit platforms), and thus not matching int.

Hence manually use int to make it work; in practice it will not be a
problem as the string with the device serial is not long.